Augmentations
There are a few client operations defined on boto3 clients that
do not come from botocore service definitions as they are
convenience functions on top of API methods and not low-level
API methods themselves. As such, an augmentations service package
has been added to manually define those methods in the same
structure as the botocore service definitions so that those can
be merged into the service definitions and treated in the same
fashion by the Lobotomy lookups. The first of these has been
added for the `s3.upload_file` client call.