Complex YAML IO
Modifies the YAML IO process such that IO only acts on the
clients and sessions needed for the lobotomy data and
preserves the rest of the file without issue. It will then
be possible to use YAML classes in the other parts of the
file and to custom format the file without losing that
formatting when adding new service calls.